ros_callbacks and pub/sub creation in main instead of ros2_node.rs. From string topics to specific blueboat_interfaces topics. Pos: u8 to f64. Not sure if can publish directly in tdma. Not tested/compiled yet

+++ src/modem_driver.rs
@@ -13,7 +13,7 @@ pub trait ModemAbstraction {
     ///   345 represents 34.5 ppt)
     fn configure(&mut self, usbl: bool, baud_rate: u32, beacon_id: u8, salinity: u16) -> Result<(), Box<dyn std::error::Error>>;
     
-    fn get_position(&mut self, t: u64) -> Result<Vec<u8>, Box<dyn Error>>;
+    fn get_position(&mut self, t: u64) -> Result<Vec<f64>, Box<dyn Error>>;
 
 
     /// Given a destination ID and data payload, sends the data via the modem.
